% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{cumulant}
\alias{cumulant}
\title{Cumulant}
\usage{
cumulant(x, j)
}
\arguments{
\item{x}{Numeric vector.
Sample data.}

\item{j}{Integer.
\code{j}th moment.
From 0 to 4.}
}
\description{
Cumulant
}
\details{
\deqn{
  K_2
  =
  \frac{n}{n - 1}
  m_2
}

\deqn{
  K_3
  =
  \frac{n^2}{ \left( n - 1 \right) \left( n - 2 \right) }
  m_3
}

\deqn{
  K_4
  =
  \frac{n^2}{ \left( n - 1 \right) \left( n - 2 \right) \left( n - 3 \right) }
  \left\{
    \left(
      n + 1
    \right)
    m_4 - 3
    \left(
      n - 1
    \right)
    m_{2}^{2}
  \right\}
}
}
\seealso{
Other moments functions: 
\code{\link{kurt}()},
\code{\link{moment}()},
\code{\link{skew}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{moments functions}
\keyword{moments}
