% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurt}
\alias{kurt}
\title{Kurtosis}
\usage{
kurt(x, type = 2, excess = TRUE)
}
\arguments{
\item{x}{Numeric vector.
Sample data.}

\item{type}{Integer.
1, 2, or 3.
See details.}

\item{excess}{Logical.
Return excess kurtosis
(kurtosis minus 3).}
}
\description{
Calculates kurtosis
which is a measure of the "taildness"
of a distribution.
}
\details{
Type 1
\deqn{
  g_2
  =
  \frac{m_4}{m_{2}^{2}}
  %(\#eq:dist-moments-g2)
}

Type 2
\deqn{
  G_2
  =
  \frac{K_4}{K_{2}^{2}}
  %(\#eq:dist-moments-G2)
}

Type 3
\deqn{
  b_2
  =
  \frac{m_4}{s^4}
  %(\#eq:dist-moments-b2)
}
}
\examples{
x <- c(
  10, 11, 12, 12, 12, 13, 13, 13, 13, 13,
  13, 14, 14, 14, 14, 15, 15, 15, 16, 17
)
kurt(x, type = 1)
kurt(x, type = 2)
kurt(x, type = 3)
x <- rnorm(n = 1000)
kurt(x, type = 1)
kurt(x, type = 2)
kurt(x, type = 3)
}
\references{
Joanes, D., & Gill, C. (1998).
Comparing Measures of Sample Skewness and Kurtosis.
\emph{Journal of the Royal Statistical Society. Series D (The Statistician)}, \emph{47}(1), 183-189.
\href{https://www.jstor.org/stable/2988433}{2988433}

\href{https://en.wikipedia.org/wiki/Kurtosis}{Wikipedia: Kurtosis}
}
\seealso{
Other moments functions: 
\code{\link{cumulant}()},
\code{\link{moments}()},
\code{\link{moment}()},
\code{\link{skew}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{moments functions}
\keyword{moments}
