% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mardia}
\alias{mardia}
\title{Multivariate Skewness and Kurtosis}
\usage{
mardia(X)
}
\arguments{
\item{X}{Matrix or data frame.}
}
\value{
Returns a vector with the following elements
\describe{
\item{b1}{Estimate of multivariate skewness \eqn{\left( b_{1, k} \right)} .}
\item{b1.chisq}{chi-square statistic \eqn{\left( \frac{nb_{1, k}}{6} \right)} .}
\item{b1.df}{Degrees of freedom \eqn{\left( \frac{k(k + 1)(k + 2)}{6} \right)} .}
\item{b1.p}{p-value associated with the chi-square statistic.}
\item{b2}{Estimate of multivariate kurtosis \eqn{\left( b_{2, k} \right)} .}
\item{b2.z}{z-statistic \eqn{\left( \frac{b_{2, k} - k \left( k + 2 \right)}{\sqrt{\frac{8 k \left( k + 2 \right)}{n}}} \right)} .}
\item{b2.p}{p-value associated with the z-statistic.}
}
}
\description{
Multivariate Skewness and Kurtosis
}
\examples{
set.seed(42)
n <- 100
mu <- c(0, 0, 0)
Sigma <- matrix(
  data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1),
  ncol = 3
)
X <- MASS::mvrnorm(n = n, mu = mu, Sigma = Sigma)
mardia(X)
}
\author{
Ivan Jacob Agaloos Pesigan
}
