% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mardiakurt}
\alias{mardiakurt}
\title{Multivariate Kurtosis}
\usage{
mardiakurt(X)
}
\arguments{
\item{X}{Matrix or data frame.}
}
\value{
Returns a vector with the following elements
\describe{
\item{b2}{Estimate of multivariate kurtosis \eqn{\left( b_{2, k} \right)} .}
\item{z}{z-statistic \eqn{\left( \frac{b_{2, k} - k \left( k + 2 \right)}{\sqrt{\frac{8 k \left( k + 2 \right)}{n}}} \right)} .}
\item{p}{p-value associated with the z-statistic.}
}
}
\description{
Mardia's estimate of multivariate kurtosis is given by
\deqn{
    b_{2, k}
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left[
      \left(
        \mathbf{X}_{i} - \mathbf{\bar{X}}
      \right)^{T}
      \boldsymbol{\hat{\Sigma}}^{-1}
      \left(
        \mathbf{X}_{i} - \mathbf{\bar{X}}
      \right)
    \right]^{2}
  }
where
\itemize{
\item \eqn{\mathbf{X}} is the \eqn{n \times k} sample data
\item \eqn{\mathbf{\bar{X}}} represent sample means
\item \eqn{\mathbf{X}_{i} - \mathbf{\bar{X}}} represent deviations from the mean
\item \eqn{\boldsymbol{\hat{\Sigma}}} is the estimated variance-covariance matrix of \eqn{\mathbf{X}} using sample data
}
}
\details{
If the null hypothesis that \eqn{\mathbf{X}} comes from a multivariate normal distribution is true,
\eqn{b_{2, k}} follows a normal distribution with a mean of \eqn{k \left( k + 2 \right)}
and a variance of \eqn{\frac{8 k \left( k + 2 \right)}{n}}.
Consequently,
\deqn{
    \frac{b_{2, k} - k \left( k + 2 \right)}{\sqrt{\frac{8 k \left( k + 2 \right)}{n}}}
  }
asymptotically follows a standard normal distribution \eqn{\mathcal{N} \left( 0, 1 \right)} .
}
\examples{
set.seed(42)
n <- 100
mu <- c(0, 0, 0)
Sigma <- matrix(
  data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1),
  ncol = 3
)
X <- MASS::mvrnorm(n = n, mu = mu, Sigma = Sigma)
mardiakurt(X)
}
\references{
Mardia, K. V. (1970).
Measures of multivariate skewness and kurtosis with applications.
\emph{Biometrika}, \emph{57}(3), 519-530.
doi:\href{https://doi.org/10.2307/2334770}{10.2307/2334770}.

Mardia, K. V. (1974).
Applications of Some Measures of Multivariate Skewness and Kurtosis in Testing Normality and Robustness Studies.
\emph{SankhyÄ: The Indian Journal of Statistics, Series B (1960-2002)}, \emph{36}(2), 115-128.
}
\author{
Ivan Jacob Agaloos Pesigan
}
