% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mardiaskew}
\alias{mardiaskew}
\title{Multivariate Skewness}
\usage{
mardiaskew(X)
}
\arguments{
\item{X}{Matrix or data frame.}
}
\value{
Returns a vector with the following elements
\describe{
\item{b1}{Estimate of multivariate skewness \eqn{\left( b_{1, k} \right)} .}
\item{chisq}{chi-square statistic \eqn{\left( \frac{nb_{1, k}}{6} \right)} .}
\item{df}{Degrees of freedom \eqn{\left( \frac{k(k + 1)(k + 2)}{6} \right)} .}
\item{p}{p-value associated with the chi-square statistic.}
}
}
\description{
Mardia's estimate of multivariate skewness is given by
\deqn{
    b_{1, k}
    =
    \frac{1}{n^2}
    \sum_{i = 1}^{n}
    \sum_{j = 1}^{n}
    \left[
      \left(
        \mathbf{X}_{i} - \mathbf{\bar{X}}
      \right)^{T}
      \boldsymbol{\hat{\Sigma}}^{-1}
      \left(
        \mathbf{X}_{j} - \mathbf{\bar{X}}
      \right)
    \right]^{3}
  }
where
\itemize{
\item \eqn{\mathbf{X}} is the \eqn{n \times k} sample data
\item \eqn{\mathbf{\bar{X}}} represent sample means
\item \eqn{\mathbf{X}_{i} - \mathbf{\bar{X}}} and \eqn{\mathbf{X}_{j} - \mathbf{\bar{X}}} represent deviations from the mean
\item \eqn{\boldsymbol{\hat{\Sigma}}} is the estimated variance-covariance matrix of \eqn{\mathbf{X}} using sample data
}
}
\details{
If the null hypothesis that \eqn{\mathbf{X}} comes from a multivariate normal distribution is true,
\eqn{\frac{nb_{1, k}}{6}} follows a chi-square \eqn{\left( \chi^2 \right)} distribution with a df of \eqn{\frac{k(k + 1)(k + 2)}{6}} .
}
\examples{
set.seed(42)
n <- 100
mu <- c(0, 0, 0)
Sigma <- matrix(
  data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1),
  ncol = 3
)
X <- MASS::mvrnorm(n = n, mu = mu, Sigma = Sigma)
mardiaskew(X)
}
\references{
Mardia, K. V. (1970).
Measures of multivariate skewness and kurtosis with applications.
\emph{Biometrika}, \emph{57}(3), 519-530.
doi:\href{https://doi.org/10.2307/2334770}{10.2307/2334770}.

Mardia, K. V. (1974).
Applications of Some Measures of Multivariate Skewness and Kurtosis in Testing Normality and Robustness Studies.
\emph{SankhyÄ: The Indian Journal of Statistics, Series B (1960-2002)}, \emph{36}(2), 115-128.
}
\author{
Ivan Jacob Agaloos Pesigan
}
