% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moment}
\alias{moment}
\title{Central Moments}
\usage{
moment(x, j)
}
\arguments{
\item{x}{Numeric vector.
Sample data.}

\item{j}{Integer.
\code{j}th moment.
From 0 to 4.}
}
\description{
The sample central moment is defined by
\deqn{
  m_j
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    x_i
    -
    \bar{x}
  \right)^j
  %(\#eq:dist-moments-sample-central)
}
where
\itemize{
\item \eqn{n} is the sample size,
\item \eqn{x = \{x_1 \dots x_n\}} is a set of values of a random variable \eqn{X},
\item \eqn{\bar{x}} is the sample mean of \eqn{x}, and
\item \eqn{j} is the \eqn{j}th central moment.
}
}
\details{
\deqn{
  m_0
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    x_i
    -
    \bar{x}
  \right)^0
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    1
  \right)
  =
  \frac{1}{n}
  n
  =
  \frac{n}{n}
  =
  1
  %(\#eq:dist-moments-sample-central-zero)
}

\deqn{
  m_1
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    x_i
    -
    \bar{x}
  \right)^1
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    0
  \right)
  =
  \frac{1}{n}
  0
  =
  \frac{0}{n}
  =
  0
  %(\#eq:dist-moments-sample-central-first)
}

\deqn{
  m_2
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    x_i
    -
    \bar{x}
  \right)^2
  %(\#eq:dist-moments-sample-central-second)
}

\deqn{
  m_3
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    x_i
    -
    \bar{x}
  \right)^3
  %(\#eq:dist-moments-sample-central-third)
}

\deqn{
  m_4
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \left(
    x_i
    -
    \bar{x}
  \right)^4
  %(\#eq:dist-moments-sample-central-fourth)
}
\itemize{
\item The "zeroth" central moment is 1.
\item The first central moment is 0.
\item The second cental moment is the variance.
\item The third central moment is used to define skewness.
\item The fourth central moment is used to define kurtosis.
}
}
\references{
\href{https://en.wikipedia.org/wiki/Central_moment}{Wikipedia: Central Moment}

\href{https://en.wikipedia.org/wiki/Standardized_moment}{Wikipedia: Standardized Moment}
}
\seealso{
Other moments functions: 
\code{\link{cumulant}()},
\code{\link{kurt}()},
\code{\link{moments}()},
\code{\link{skew}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{moments functions}
\keyword{moments}
