% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Compute mean, variance, standard deviation, skewness, and kurtosis}
\usage{
moments(x, type = 2)
}
\arguments{
\item{x}{Numeric vector.
Sample data.}

\item{type}{Integer.
1, 2, or 3.
See \code{\link[=skew]{skew()}} and \code{\link[=kurt]{kurt()}}.}
}
\description{
Compute mean, variance, standard deviation, skewness, and kurtosis
}
\seealso{
Other moments functions: 
\code{\link{cumulant}()},
\code{\link{kurt}()},
\code{\link{moment}()},
\code{\link{skew}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{moments functions}
\keyword{moments}
