% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn2ll}
\alias{mvn2ll}
\title{Multivariate Normal - Two Log-Likelihood}
\usage{
mvn2ll(X, mu, Sigma, neg = TRUE)
}
\arguments{
\item{X}{Numeric matrix.
Values of the \eqn{k}-dimensional random variable \eqn{\mathbf{X}}.}

\item{mu}{Numeric vector.
Location parameter mean vector \eqn{\boldsymbol{\mu}} of length \eqn{k}.}

\item{Sigma}{Numeric matrix.
\eqn{k \times k} variance-covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{neg}{Logical.
If \code{TRUE},
returns,
negative two log-likelihood.}
}
\description{
Calculates the two log-likelihood of \eqn{\mathbf{X}}
following a mutivariate normal distribution.
}
\details{
The two log-likelihood for the multivariate normal
(or multivariate Gaussian,
or joint normal)
distribution is given by
\deqn{
  2
  \ln \mathcal{L}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  2
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right) \\
  2
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  -
  \ln
  \left(
    | \boldsymbol{\Sigma} |
  \right)
  -
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)^{\prime}
  \boldsymbol{\Sigma}^{-1}
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)
  -
  k \ln
  \left(
  2 \pi
  \right)
  %(\#eq:dist-mvn2ll)
}
with
\eqn{k}-dimensional
random vector
\eqn{\mathbf{X} \in \boldsymbol{\mu} + \textrm{span}\left(\boldsymbol{\Sigma}\right) \subseteq \mathbf{R}^k},
\eqn{\boldsymbol{\mu}}
is the location parameter mean
\eqn{\left( \boldsymbol{\mu} \in \mathbf{R}^k \right)},
and
\eqn{\boldsymbol{\Sigma}}
is the variance-covariance matrix
\eqn{\left( \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} \right)}.

The negative two log-likelihood is given by
\deqn{
  -2
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  \ln
  \left(
    | \boldsymbol{\Sigma} |
  \right)
  +
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)^{\prime}
  \boldsymbol{\Sigma}^{-1}
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)
  +
  k \ln
  \left(
  2 \pi
  \right) .
  %(\#eq:dist-mvnneg2ll)
}
}
\references{
\href{https://en.wikipedia.org/wiki/Multivariate_normal_distribution}{Wikipedia: Multivariate Normal Distribution}
}
\seealso{
Other mutivariate normal likelihood functions: 
\code{\link{mvnll}()},
\code{\link{mvnpdf}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{mutivariate normal likelihood functions}
\keyword{mutivariate}
\keyword{normal}
