% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal.R
\name{mvnormal_2ll}
\alias{mvnormal_2ll}
\title{Multivariate Normal - Two Log-Likelihood}
\usage{
mvnormal_2ll(X, mu, Sigma, neg = TRUE)
}
\arguments{
\item{X}{Numeric matrix.
Values of the \eqn{k}-dimensional random variable \eqn{X}.}

\item{mu}{Numeric vector.
Location parameter mean vector of length \eqn{k}
(\eqn{\boldsymbol{\mu}}).}

\item{Sigma}{Numeric matrix.
\eqn{k \times k} variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}

\item{neg}{Logical.
If \code{TRUE},
returns,
negative two log-likelihood.}
}
\description{
Calculates the two log-likelihood of \eqn{\mathbf{X}}
following a mutivariate normal distribution.
}
\details{
The two log-likelihood for the multivariate normal
(or multivariate Gaussian,
or joint normal)
distribution is given by
\deqn{
  2
  \ln \mathcal{L}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  2
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right) \\
  2
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  -
  \ln
  \left(
    | \boldsymbol{\Sigma} |
  \right)
  -
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)^{\prime}
  \boldsymbol{\Sigma}^{-1}
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)
  -
  k \ln
  \left(
  2 \pi
  \right)
}
with
\eqn{k}-dimensional
random vector
\eqn{\mathbf{X} \in \boldsymbol{\mu} + \textrm{span}\left(\boldsymbol{\Sigma}\right) \subseteq \mathbf{R}^k},
\eqn{\boldsymbol{\mu}}
is the location parameter mean
(\eqn{\boldsymbol{\mu} \in \mathbf{R}^k}),
and
\eqn{\boldsymbol{\Sigma}}
is the variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \in \mathbf{R}^{k \times k}}).

The negative two log-likelihood is given by
\deqn{
  -2
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  \ln
  \left(
    | \boldsymbol{\Sigma} |
  \right)
  +
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)^{\prime}
  \boldsymbol{\Sigma}^{-1}
  \left(
    \mathbf{x} - \boldsymbol{\mu}
  \right)
  +
  k \ln
  \left(
  2 \pi
  \right) .
}
}
\references{
\href{https://en.wikipedia.org/wiki/Multivariate_normal_distribution}{Wikipedia: Multivariate Normal Distribution}
}
\author{
Ivan Jacob Agaloos Pesigan
}
