% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Multivariate Normal - Probablity Density Function}
\usage{
mvnpdf(X, mu, Sigma)
}
\arguments{
\item{X}{Numeric matrix.
Values of the \eqn{k}-dimensional random variable \eqn{\mathbf{X}}.}

\item{mu}{Numeric vector.
Location parameter mean vector \eqn{\boldsymbol{\mu}} of length \eqn{k}.}

\item{Sigma}{Numeric matrix.
\eqn{k \times k} variance-covariance matrix
\eqn{\boldsymbol{\Sigma}}.}
}
\description{
Calculates probablities
from the probability density function
of the multivariate normal distribution
\eqn{
  \mathbf{X}
  \sim
  \mathcal{N}_{k}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
  \right) .
  %(\#eq:dist-X-mvn)
}
}
\details{
The multivariate normal
(or multivariate Gaussian,
or joint normal)
distribution is given by
\deqn{
  \mathbf{X}
  \sim
  \mathcal{N}_{k}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
  \right)
  %(\#eq:dist-X-mvn)
}
and has the probability density function (PDF)
\deqn{
  f_{\mathbf{X}}
  \left(
    X_1,
    \dots,
    X_k
  \right)
  =
  \frac{
    \exp
    \left[
      -
      \frac{1}{2}
      \left(
        \mathbf{X} - \boldsymbol{\mu}
      \right)^{\prime}
      \boldsymbol{\Sigma}^{-1}
      \left(
        \mathbf{X} - \boldsymbol{\mu}
      \right)
    \right]
  }{
    \sqrt{
      \left(
        2 \pi
      \right)^{k}
      | \boldsymbol{\Sigma} |
    }
  }
  %(\#eq:dist-mvnpdf)
}
with
\eqn{k}-dimensional
random vector
\eqn{\mathbf{X} \in \boldsymbol{\mu} + \textrm{span}\left(\boldsymbol{\Sigma}\right) \subseteq \mathbf{R}^k},
\eqn{\boldsymbol{\mu}}
is the location parameter mean
\eqn{\left( \boldsymbol{\mu} \in \mathbf{R}^k \right)},
and
\eqn{\boldsymbol{\Sigma}}
is the variance-covariance matrix
\eqn{\left( \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} \right)}.
}
\references{
\href{https://en.wikipedia.org/wiki/Multivariate_normal_distribution}{Wikipedia: Multivariate Normal Distribution}
}
\seealso{
Other mutivariate normal likelihood functions: 
\code{\link{mvn2ll}()},
\code{\link{mvnll}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{mutivariate normal likelihood functions}
\keyword{mutivariate}
\keyword{normal}
