% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_2ll}
\alias{normal_2ll}
\title{Normal - Two Log-Likelihood}
\usage{
normal_2ll(mu, sigma, x, neg = TRUE)
}
\arguments{
\item{mu}{Numeric.
Location parameter mean (\eqn{\mu}).}

\item{sigma}{Numeric.
Positive number.
Scale parameter standard deviation (\eqn{\sigma = \sqrt{\sigma^2}}).}

\item{x}{Numeric vector.
Values of the random variable \eqn{X}.}

\item{neg}{Logical.
If \code{TRUE},
returns,
negative two log-likelihood.}
}
\description{
Calculates the two log-likelihood of \eqn{X}
following a normal distribution.
}
\details{
The two log-likelihood for the normal (or Gaussian or Gauss or Laplace–Gauss)
distribution is given by
\deqn{
    2
    \mathcal{l}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    2
    \left[
      -
      \frac{n}{2}
      \ln 2 \pi \sigma^2
      -
      \frac{1}{2 \sigma^2}
      \sum_{i = 1}^{n}
      \left(
        x_i - \mu
      \right)^2
    \right] \\
    =
    -
    n
    \ln 2 \pi \sigma^2
    -
    \frac{1}{\sigma^2}
    \sum_{i = 1}^{n}
    \left(
      x_i - \mu
    \right)^2
  }
with
independent and identically distributed
sample data \eqn{x \in \mathbb{R}},
\eqn{\mu}
is the location parameter mean being estimated
(\eqn{\mu \in \mathbb{R}}),
and
\eqn{\sigma^2}
is the scale parameter variance being estimated
(\eqn{\sigma^2 > 0}).

The negative two log-likelihood is given by
\deqn{
    -
    2
    \mathcal{l}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    -
    2
    \left[
      -
      \frac{n}{2}
      \ln 2 \pi \sigma^2
      -
      \frac{1}{2 \sigma^2}
      \sum_{i = 1}^{n}
      \left(
        x_i - \mu
      \right)^2
    \right] \\
    =
    n
    \ln 2 \pi \sigma^2
    +
    \frac{1}{\sigma^2}
    \sum_{i = 1}^{n}
    \left(
      x_i - \mu
    \right)^2 .
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Normal_distribution}{Wikipedia: Normal Distribution}

\href{https://en.wikipedia.org/wiki/Independent_and_identically_distributed_random_variables}{Wikipedia: Independent and Identically Distributed Random Variables}

\href{https://en.wikipedia.org/wiki/Likelihood_function}{Wikipedia: Likelihood Function}
}
\seealso{
Other normal likelihood functions: 
\code{\link{normal_L}()},
\code{\link{normal_ll}()},
\code{\link{normal_obj}()}
}
\concept{normal likelihood functions}
