% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_L}
\alias{normal_L}
\title{Normal - Likelihood}
\usage{
normal_L(mu, sigma, x)
}
\arguments{
\item{mu}{Numeric.
Location parameter mean (\eqn{\mu}).}

\item{sigma}{Numeric.
Positive number.
Scale parameter standard deviation (\eqn{\sigma = \sqrt{\sigma^2}}).}

\item{x}{Numeric vector.
Values of the random variable \eqn{X}.}
}
\description{
Calculates the likelihood of \eqn{X}
following a normal distribution.
}
\details{
The likelihood function for the normal (or Gaussian or Gauss or Laplace–Gauss)
distribution is given by
\deqn{
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \prod_{i = 1}^{n}
    \left\{
      \frac{1}{\sigma \sqrt{2 \pi}}
      \exp
        \left[
          -
          \frac{1}{2}
          \left(
            \frac{x_i - \mu}{\sigma}
          \right)^2
        \right]
    \right\}
  }
or
\deqn{
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \prod_{i = 1}^{n}
    \left\{
      \frac{1}{\sqrt{2 \pi \sigma^2}}
      \exp
        \left[
          -
          \frac{
          \left(
            x_i - \mu
          \right)^2}
          {2 \sigma^2}
        \right]
    \right\} \\
    =
    \left(
      \frac{1}{\sqrt{2 \pi \sigma^2}}
    \right)^n
    \exp
      \left[
        -
        \frac{1}{2 \sigma^2}
        \sum_{i = 1}^{n}
        \left(
          x_i - \mu
        \right)^2
      \right]
    }
with
independent and identically distributed
sample data \eqn{x \in \mathbb{R}},
\eqn{\mu}
is the location parameter mean being estimated
(\eqn{\mu \in \mathbb{R}}),
and
\eqn{\sigma^2}
is the scale parameter variance being estimated
(\eqn{\sigma^2 > 0}).
}
\references{
\href{https://en.wikipedia.org/wiki/Normal_distribution}{Wikipedia: Normal Distribution}

\href{https://en.wikipedia.org/wiki/Independent_and_identically_distributed_random_variables}{Wikipedia: Independent and Identically Distributed Random Variables}

\href{https://en.wikipedia.org/wiki/Likelihood_function}{Wikipedia: Likelihood Function}
}
\seealso{
Other normal likelihood functions: 
\code{\link{normal_2ll}()},
\code{\link{normal_ll}()},
\code{\link{normal_obj}()}
}
\concept{normal likelihood functions}
