% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_negll}
\alias{normal_negll}
\title{Normal - Negative Log-Likelihood}
\usage{
normal_negll(theta, x)
}
\arguments{
\item{theta}{Vector of parameters \eqn{\theta} of the normal distribution
(\code{theta[1] = mu} (\eqn{\mu}) and \code{theta[2] = sigma} (\eqn{\sigma})).}

\item{x}{Numeric vector.
Independent and identically distributed sample data.}
}
\description{
Calculates the negative log-likelihood of \eqn{x}
following a normal distribution.
This function is identical to
\code{-sum(dnorm(x = x, mean = theta[1], sd = theta[2], log = TRUE))}.
}
\details{
The likelihood function for the normal (or Gaussian or Gauss or Laplace–Gauss)
distribution is given by
\deqn{
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \prod_{i = 1}^{n}
    \left\{
      \frac{1}{\sigma \sqrt{2 \pi}}
      \exp
        \left[
          -
          \frac{1}{2}
          \left(
            \frac{x_i - \mu}{\sigma}
          \right)^2
        \right]
    \right\}
  }
or
\deqn{
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \prod_{i = 1}^{n}
    \left\{
      \frac{1}{\sqrt{2 \pi \sigma^2}}
      \exp
        \left[
          -
          \frac{
          \left(
            x_i - \mu
          \right)^2}
          {2 \sigma^2}
        \right]
    \right\} \\
    =
    \left(
      \frac{1}{\sqrt{2 \pi \sigma^2}}
    \right)^n
    \exp
      \left[
        -
        \frac{1}{2 \sigma^2}
        \sum_{i = 1}^{n}
        \left(
          x_i - \mu
        \right)^2
      \right]
    }
with
independent and identically distributed
sample data \eqn{x \in \mathbb{R}},
\eqn{\mu}
is the location parameter mean being estimated
(\eqn{\mu \in \mathbb{R}}),
and
\eqn{\sigma^2}
is the scale parameter variance being estimated
(\eqn{\sigma^2 > 0}).

The natural log of the likelihood function is given by
\deqn{
    \ln
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    -
    \frac{n}{2}
    \ln 2 \pi \sigma^2
    -
    \frac{1}{2 \sigma^2}
    \sum_{i = 1}^{n}
    \left(
      x_i - \mu
    \right)^2 .
  }

The negative log-likelihood is given by
\deqn{
    -
    \ln
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \frac{n}{2}
    \ln 2 \pi \sigma^2
    +
    \frac{1}{2 \sigma^2}
    \sum_{i = 1}^{n}
    \left(
      x_i - \mu
    \right)^2 .
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Normal_distribution}{Wikipedia: Normal Distribution}

\href{https://en.wikipedia.org/wiki/Independent_and_identically_distributed_random_variables}{Wikipedia: Independent and Identically Distributed Random Variables}

\href{https://en.wikipedia.org/wiki/Likelihood_function}{Wikipedia: Likelihood Function}
}
\author{
Ivan Jacob Agaloos Pesigan
}
