% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_obj}
\alias{normal_obj}
\title{Normal Distribution - Objective Function}
\usage{
normal_obj(theta, x, neg = TRUE)
}
\arguments{
\item{theta}{Vector of parameters \eqn{\theta} of the normal distribution
(\code{theta[1] = mu} (\eqn{\mu}) and \code{theta[2] = sigma} (\eqn{\sigma})).}

\item{x}{Numeric vector.
Values of the random variable \eqn{X}.}

\item{neg}{Logical.
If \code{TRUE},
returns,
negative log-likelihood.}
}
\description{
Objective function to minimize/maximize
to estimate parameters of the normal distribution.
See \code{\link[=normal_2ll]{normal_2ll()}}.
}
\references{
\href{https://en.wikipedia.org/wiki/Normal_distribution}{Wikipedia: Normal Distribution}

\href{https://en.wikipedia.org/wiki/Independent_and_identically_distributed_random_variables}{Wikipedia: Independent and Identically Distributed Random Variables}

\href{https://en.wikipedia.org/wiki/Likelihood_function}{Wikipedia: Likelihood Function}
}
\seealso{
Other normal likelihood functions: 
\code{\link{normal_2ll}()},
\code{\link{normal_L}()},
\code{\link{normal_ll}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{normal likelihood functions}
