% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_pdf}
\alias{normal_pdf}
\title{Normal Distribution - Probablity Density Function}
\usage{
normal_pdf(x, mu = 0, sigma = 1, log = FALSE)
}
\arguments{
\item{x}{Numeric vector.
Values of the random variable \eqn{X}.}

\item{mu}{Numeric.
Location parameter mean (\eqn{\mu}).}

\item{sigma}{Numeric.
Positive number.
Scale parameter standard deviation (\eqn{\sigma = \sqrt{\sigma^2}}).}

\item{log}{Logical.
If \code{TRUE},
returns \eqn{\log \left( f \left( x \right) \right)}.}
}
\value{
Returns \eqn{f \left( x \right)}
using the probablity density function
with the supplied parameter/s.
If \code{log = TRUE},
returns \eqn{\log \left( f \left( x \right) \right)}.
}
\description{
Calculates probablities
from the probability density function
of the normal distribution.
This function is identical to \code{\link[=dnorm]{dnorm()}}.
}
\details{
The normal (or Gaussian or Gauss or Laplace–Gauss)
distribution is given by
\deqn{
    \mathcal{N}
      \left(
        \mu,
        \sigma^2
      \right)
  }
and has
the probability density function (PDF)
\deqn{
    f \left( x \right)
    =
    \frac{1}{\sigma \sqrt{2 \pi}}
    \exp
      \left[
        -
        \frac{1}{2}
        \left(
          \frac{x - \mu}{\sigma}
        \right)^2
      \right]
  }
or
\deqn{
    f \left( x \right)
    =
    \frac{1}{\sqrt{2 \pi \sigma^2}}
    \exp
      \left[
        -
        \frac{
        \left(
          x - \mu
        \right)^2}
        {2 \sigma^2}
      \right]
  }
with
\eqn{x \in \mathbf{R}},
\eqn{\mu}
is the location parameter mean
(\eqn{\mu \in \mathbf{R}}),
and
\eqn{\sigma^2}
is the scale parameter variance
(\eqn{\sigma^2 > 0}).
}
\references{
\href{https://en.wikipedia.org/wiki/Normal_distribution}{Wikipedia: Normal Distribution}
}
\author{
Ivan Jacob Agaloos Pesigan
}
