% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{normll}
\alias{normll}
\title{Normal - Log-Likelihood}
\usage{
normll(mu, sigma, x, neg = TRUE)
}
\arguments{
\item{mu}{Numeric.
Location parameter mean (\eqn{\mu}).}

\item{sigma}{Numeric.
Positive number.
Scale parameter standard deviation (\eqn{\sigma = \sqrt{\sigma^2}}).}

\item{x}{Numeric vector.
Values of the random variable \eqn{X}.}

\item{neg}{Logical.
If \code{TRUE},
returns,
negative log-likelihood.}
}
\description{
Calculates the log-likelihood of \eqn{X}
following a normal distribution.
}
\details{
The natural log of the likelihood function for the normal (or Gaussian or Gauss or Laplace–Gauss)
distribution is given by
\deqn{
    \ln
    \mathcal{L}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \mathcal{l}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right) \\
    \mathcal{l}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    \ln
    \left\{
      \left(
        \frac{1}{\sqrt{2 \pi \sigma^2}}
      \right)^n
      \exp
        \left[
          -
          \frac{1}{2 \sigma^2}
          \sum_{i = 1}^{n}
          \left(
            x_i - \mu
          \right)^2
        \right]
    \right\} \\
    =
    -
    \frac{n}{2}
    \ln 2 \pi \sigma^2
    -
    \frac{1}{2 \sigma^2}
    \sum_{i = 1}^{n}
    \left(
      x_i - \mu
    \right)^2 .
  }
with
independent and identically distributed
sample data \eqn{x \in \mathbf{R}},
\eqn{\mu}
is the location parameter mean being estimated
(\eqn{\mu \in \mathbf{R}}),
and
\eqn{\sigma^2}
is the scale parameter variance being estimated
(\eqn{\sigma^2 > 0}).

The negative log-likelihood is given by
\deqn{
    -
    \mathcal{l}
      \left(
        \mu,
        \sigma^2
        \mid
        x
      \right)
    =
    -
    \left[
      -
      \frac{n}{2}
      \ln 2 \pi \sigma^2
      -
      \frac{1}{2 \sigma^2}
      \sum_{i = 1}^{n}
      \left(
        x_i - \mu
      \right)^2
    \right] \\
    =
    \frac{n}{2}
    \ln 2 \pi \sigma^2
    +
    \frac{1}{2 \sigma^2}
    \sum_{i = 1}^{n}
    \left(
      x_i - \mu
    \right)^2 .
  }
}
\references{
\href{https://en.wikipedia.org/wiki/normaldistribution}{Wikipedia: Normal Distribution}

\href{https://en.wikipedia.org/wiki/Independent_and_identically_distributed_random_variables}{Wikipedia: IID}

\href{https://en.wikipedia.org/wiki/Likelihood_function}{Wikipedia: Likelihood Function}
}
\seealso{
Other normal likelihood functions: 
\code{\link{norm2ll}()},
\code{\link{normL}()},
\code{\link{normobj}()}
}
\concept{normal likelihood functions}
