% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skew}
\alias{skew}
\title{Skewness}
\usage{
skew(x, type = 2)
}
\arguments{
\item{x}{Numeric vector.
Sample data.}

\item{type}{Integer.
1, 2, or 3.
See details.}
}
\description{
Calculates kurtosis
which is a measure of the asymmetry
of a distribution.
}
\details{
Type 1
\deqn{
  g_1
  =
  \frac{m_3}{m_{2}^{\frac{3}{2}}}
  %(\#eq:dist-moments-g1)
}

Type 2
\deqn{
  G_1
  =
  \frac{K_3}{K_{2}^{\frac{3}{2}}}
  %(\#eq:dist-moments-G1)
}

Type 3
\deqn{
  b_1
  =
  \frac{m_3}{s^3}
  %(\#eq:dist-moments-b1)
}
}
\examples{
x <- c(
  10, 11, 12, 12, 12, 13, 13, 13, 13, 13,
  13, 14, 14, 14, 14, 15, 15, 15, 16, 17
)
skew(x, type = 1)
skew(x, type = 2)
skew(x, type = 3)
x <- rnorm(n = 1000)
skew(x, type = 1)
skew(x, type = 2)
skew(x, type = 3)
}
\references{
Joanes, D., & Gill, C. (1998).
Comparing Measures of Sample Skewness and Kurtosis.
\emph{Journal of the Royal Statistical Society. Series D (The Statistician)}, \emph{47}(1), 183-189.
\href{https://www.jstor.org/stable/2988433}{2988433}

\href{https://en.wikipedia.org/wiki/Skewness}{Wikipedia: Kurtosis}
}
\seealso{
Other moments functions: 
\code{\link{cumulant}()},
\code{\link{kurt}()},
\code{\link{moments}()},
\code{\link{moment}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{moments functions}
\keyword{moments}
