% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvnll}
\alias{mvnll}
\title{Multivariate Normal - Log-Likelihood}
\usage{
mvnll(X, mu, Sigma, neg = TRUE)
}
\arguments{
\item{X}{Numeric matrix.
Values of the \eqn{k}-dimensional random variable \eqn{X}.}

\item{mu}{Numeric vector.
Location parameter mean vector of length \eqn{k}
(\eqn{\boldsymbol{\mu}}).}

\item{Sigma}{Numeric matrix.
\eqn{k \times k} variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}

\item{neg}{Logical.
If \code{TRUE},
returns,
negative log-likelihood.}
}
\description{
Calculates the log-likelihood of \eqn{\mathbf{X}}
following a mutivariate normal distribution.
}
\details{
The natural log of the likelihood function for the multivariate normal
(or multivariate Gaussian,
or joint normal)
distribution is given by
\deqn{
  \ln \mathcal{L}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right) \\
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  -
  \frac{1}{2}
  \left[
    \ln
    \left(
      | \boldsymbol{\Sigma} |
    \right)
    +
    \left(
      \mathbf{x} - \boldsymbol{\mu}
    \right)^{\prime}
    \boldsymbol{\Sigma}^{-1}
    \left(
      \mathbf{x} - \boldsymbol{\mu}
    \right)
    +
    k \ln
    \left(
    2 \pi
    \right)
  \right]
}
with
\eqn{k}-dimensional
random vector
\eqn{\mathbf{X} \in \boldsymbol{\mu} + \textrm{span}\left(\boldsymbol{\Sigma}\right) \subseteq \mathbf{R}^k},
\eqn{\boldsymbol{\mu}}
is the location parameter mean
(\eqn{\boldsymbol{\mu} \in \mathbf{R}^k}),
and
\eqn{\boldsymbol{\Sigma}}
is the variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \in \mathbf{R}^{k \times k}}).

The negative log-likelihood is given by
\deqn{
  -
  \mathcal{l}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
    \mid
    \mathbf{X}
  \right)
  =
  \frac{1}{2}
  \left[
    \ln
    \left(
      | \boldsymbol{\Sigma} |
    \right)
    +
    \left(
      \mathbf{x} - \boldsymbol{\mu}
    \right)^{\prime}
    \boldsymbol{\Sigma}^{-1}
    \left(
      \mathbf{x} - \boldsymbol{\mu}
    \right)
    +
    k \ln
    \left(
    2 \pi
    \right)
  \right] .
}
}
\references{
\href{https://en.wikipedia.org/wiki/Multivariate_normal_distribution}{Wikipedia: Multivariate Normal Distribution}
}
\author{
Ivan Jacob Agaloos Pesigan
}
